Fix the location of Hazelcast server and client.
