+++ modules/datacafe-server/src/main/java/edu/emory/bmi/datacafe/hazelcast/HzServer.java
@@ -13,45 +13,39 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package edu.emory.bmi.datacafe.client.samples;
+package edu.emory.bmi.datacafe.hazelcast;
 
 import edu.emory.bmi.datacafe.core.hazelcast.HzInstance;
-import org.apache.logging.log4j.LogManager;
-import org.apache.logging.log4j.Logger;
 
 import java.util.concurrent.ConcurrentMap;
 
 /**
- * The client for Hazelcast In-Memory Data Grid
+ * A simple Hazelcast Server instance
  */
-public class HzClient extends HzInstance {
-    private static Logger logger = LogManager.getLogger(HzClient.class.getName());
+public class HzServer extends HzInstance{
 
     public static void main(String[] args) {
         init();
     }
 
     /**
-     * Reads an entry from the map
-     * invoke: HzClient.readValues("my-distributed-map", "sample-key");
-     *
+     * Adds an entry to a map
+     * invoke: HzServer.addValueToMap("my-distributed-map", "sample-key", "sample-value");
      * @param mapName the name of the map
-     * @param key     the key
-     * @return the value of the entry.
+     * @param key the key
+     * @param value the value
      */
-    public static String readValues(String mapName, String key) {
+    public static void addValueToMap(String mapName, String key, String value) {
         ConcurrentMap<String, String> map = firstInstance.getMap(mapName);
-        return map.get(key);
+        map.put(key, value);
     }
 
     /**
-     * Reads and prints a value from a map.
-     *
+     * Gets a Hazelcast distributed map.
      * @param mapName the name of the map
-     * @param key     the key
+     * @return the concurrent map.
      */
-    public static void printValues(String mapName, String key) {
-        String val = readValues(mapName, key);
-        logger.info("The value is: " + val);
+    public static ConcurrentMap<String, String> getMap(String mapName) {
+        return firstInstance.getMap(mapName);
     }
 }
